Refactor cookie consent component
- Extracted types to `types.ts`
- Extracted analytics utility functions to `analytics.ts`
- Extracted custom hook `useCookieConsent` to `use-cookie-consent.ts`
- Extracted UI components `CookieBanner` and `CookiePreferencesModal` to `cookie-banner.tsx` and `cookie-preferences-modal.tsx` respectively
- Rewrote `index.tsx` to serve as the entry point, coordinating the custom hook, banner, and modal components.

This improves maintainability and readability by breaking down the monolithic >500 line component into smaller, more manageable pieces while preserving all existing functionality.
