+++ src/components/cookie-consent/cookie-banner.tsx
@@ -1,0 +1,59 @@
+import Link from 'next/link'
+
+interface CookieBannerProps {
+  onDeclineAll: () => void
+  onShowPreferences: () => void
+  onAcceptAll: () => void
+}
+
+export function CookieBanner({ onDeclineAll, onShowPreferences, onAcceptAll }: CookieBannerProps) {
+  return (
+    <div
+      className="fixed bottom-0 left-0 right-0 z-50 bg-white border-t-2 border-gray-200 shadow-2xl"
+      role="region"
+      aria-label="Cookie consent notice"
+    >
+      <div className="max-w-7xl mx-auto p-4 sm:p-6">
+        <div className="flex flex-col md:flex-row items-start md:items-center justify-between gap-4">
+          <div className="flex-1">
+            <h3 className="text-lg font-bold text-gray-900 mb-2">We Value Your Privacy</h3>
+            <p className="text-sm text-gray-600 mb-3">
+              We use cookies to improve your experience on our site, analyze traffic, and enable
+              certain features. By clicking &quot;Accept All&quot;, you consent to our use of
+              cookies for analytics and marketing purposes. You can manage your preferences or
+              decline non-essential cookies.
+            </p>
+            <div className="flex items-center gap-4 text-xs text-gray-500">
+              <Link href="/privacy-policy" className="text-blue-600 underline">
+                Privacy Policy
+              </Link>
+              <Link href="/cookie-policy" className="text-blue-600 underline">
+                Cookie Policy
+              </Link>
+            </div>
+          </div>
+          <div className="flex flex-col sm:flex-row gap-2 w-full md:w-auto">
+            <button
+              onClick={onDeclineAll}
+              className="px-6 py-2.5 bg-gray-200 text-gray-700 rounded-lg font-semibold hover:bg-gray-300 transition-colors text-sm whitespace-nowrap"
+            >
+              Decline All
+            </button>
+            <button
+              onClick={onShowPreferences}
+              className="px-6 py-2.5 bg-gray-200 text-gray-700 rounded-lg font-semibold hover:bg-gray-300 transition-colors text-sm whitespace-nowrap"
+            >
+              Customize
+            </button>
+            <button
+              onClick={onAcceptAll}
+              className="px-6 py-2.5 bg-blue-600 text-white rounded-lg font-semibold hover:bg-blue-700 transition-colors text-sm whitespace-nowrap"
+            >
+              Accept All
+            </button>
+          </div>
+        </div>
+      </div>
+    </div>
+  )
+}
